added clickable and partial images
I created a class called clickable which is the new superclass for the button class.
This allows us to make a puzzlepiece class which is also clickable and can have different actions based on the click action set in the setonclicklistener method.
The method of implementing this is demonstrated in the code.